requirements.txt
python-telegram-bot==13.15
requests==2.20.1
python-dotenv==0.9.1